+++ SIL.WritingSystems/IetfLanguageTag.cs
@@ -23,6 +23,9 @@ namespace SIL.WritingSystems
 	{
 		private const string kInvalidTagMsg = "The IETF language tag is invalid.";
 
+		private const string kSimplifiedChineseNameInEnglish = "Chinese (Simplified)";
+		private const string kTraditionalChineseNameInEnglish = "Chinese (Traditional)";
+
 		private const string PrivateUseExpr = "[xX](-" + PrivateUseSubExpr + ")+";
 		// according to RFC-5646 the private use subtag can be up to 8 characters
 		// some data in alltags uses longer ones so relaxing this requirement
@@ -1298,7 +1301,7 @@ namespace SIL.WritingSystems
 
 			if (generalCode == ChineseSimplifiedTag && uiLanguageCode == "en")
 			{
-				// This corresponds to what we (currently) get as the "English Subtitle" in
+				// This corresponds to what we use as the "English Subtitle" in
 				// GetNativeLanguageNameWithEnglishSubtitle. Not sure if it really matters here,
 				// but the ICU-supplied name (e.g., used on Linux), and the EnglishName and
 				// DisplayName supplied via the Windows CultureInfo are all subtly different in
@@ -1306,11 +1309,11 @@ namespace SIL.WritingSystems
 				// ICU: Chinese (China)
 				// DisplayName: Chinese (Simplified, PRC)
 				// EnglishName: Chinese (Simplified, China)
-				// Ideally, we should probably either have GetNativeLanguageNameWithEnglishSubtitle
-				// use this same constant or factor out the complex logic so that we always compute
-				// the same value based on a single well-defined source. But it's not clear whether
-				// one of those approaches is superior.
-				return "Chinese (Simplified)";
+
+				return kSimplifiedChineseNameInEnglish;
+			} else if (generalCode == ChineseTraditionalTag && uiLanguageCode == "en")
+			{
+				return kTraditionalChineseNameInEnglish;
 			}
 
 			// Starting some time around Sept 2025, Windows started returning the "fa" culture
@@ -1427,7 +1430,14 @@ namespace SIL.WritingSystems
 				if (IsNullOrWhiteSpace(nativeName))
 					nativeName = code;
 
-				if (ci.Name != ChineseSimplifiedTag && ci.Name != ChineseTraditionalTag)
+				if (ci.Name == ChineseSimplifiedTag)
+				{
+					englishNameSuffix = $" ({kSimplifiedChineseNameInEnglish})";
+				}
+				else if (ci.Name == ChineseTraditionalTag)
+				{
+					englishNameSuffix = $" ({kTraditionalChineseNameInEnglish})";
+				} else
 				{
 					// Remove any country (or script?) names.
 					var idxCountry = englishNameSuffix.LastIndexOf(" (", StringComparison.Ordinal);
@@ -1436,18 +1446,6 @@ namespace SIL.WritingSystems
 					idxCountry = nativeName.IndexOf(" (", StringComparison.Ordinal);
 					if (idxCountry > 0)
 						nativeName = nativeName.Substring(0, idxCountry);
-				}
-				else if (englishNameSuffix.Length > 0)
-				{
-					// I have seen more cruft after the country name a few times, so remove that
-					// as well. The parenthetical expansion always seems to start "(Simplified",
-					// which we want to keep. We need double close parentheses because there's one
-					// open parenthesis before "Chinese" and another open parenthesis before
-					// "Simplified" (which precedes ", China" or ", PRC"). Also, we don't worry
-					// about the parenthetical content of the native Chinese name.
-					var idxCountry = englishNameSuffix.IndexOf(", ", StringComparison.Ordinal);
-					if (idxCountry > 0)
-						englishNameSuffix = englishNameSuffix.Substring(0, idxCountry) + "))";
 				}
 				langName = nativeName + englishNameSuffix;
 				if (!ci.IsUnknownCulture())